viegtor/backend/requirements.txt
# API
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.2
pydantic-settings==2.6.1

# Google Cloud
google-cloud-firestore==2.19.0
google-cloud-aiplatform==1.73.0

# Scraping
playwright==1.49.0
playwright-stealth==1.0.6

# Scheduling
apscheduler==3.10.4

# HTTP / parsing
httpx==0.28.0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0

# Data sources
sec-api==1.0.26
firecrawl-py==1.4.0

# Utils
python-dotenv==1.0.1
